requirements.txt
flask
groq
python-dotenv
reportlab
pypdf2
python-docx